Revert Network to main.py, delete useless file & fix bugs
Fixes #19
