+++ classes/__init__.py
@@ -4,7 +4,6 @@ from . import blocks
 # Import files
 from . import BasicClasses
 from . import Exceptions
-from . import Network
 # Library not found, uncomment next line when this will be fixed
 #from . import PathFinder
 from . import TerrainFeatures
